Update Esp32_coop_
Moved all functions from Loop to Setup so I could add an "esp_deep_sleep" to save battery life.  
Now the esp32 will go to sleep for a set amount of time after checking the Google sheet and performing the desired operation.  
Since I am running this of a battery and solar panel setup, it was necessary to ensure the battery lasted through the night. 

I also added a safety measure in the "motorUp" and "motorDown" to turn off the motor in the event that a reed switch failed to read.  
This is in the form of a while true loop that is entered when the motor has been running for more than 20 seconds. 

Also, sped up the motors to max speed.